Webpack ignore main from renderer and vice versa, moved network connectors accordingly
